+++ tsup.config.ts
@@ -2,9 +2,10 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   entry: ["src/index.ts"],
+  injectStyle: true,
   format: ["esm"], // Build for commonJS and ESmodules
   dts: true, // Generate declaration file (.d.ts)
   splitting: false,
   sourcemap: true,
-  clean: true,
+  clean: true
 });