+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.0.6
+
+- Fix an issue where a heading could be nested inside a paragraph, which is not allowed.
+- Fix an issue when removing lists.
+
+## 1.0.5
+
+- Try to fix isHotKey ESM issue
+
+## 1.0.3 / 1.0.4
+
+- Try to fix bundling issues
+
 ## 1.0.2
 
 - Fix the package.json